@@ -21,7 +21,7 @@
         "symfony/cache-implementation": "1.1|2.0|3.0"
     },
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
         "psr/cache": "^2.0|^3.0",
         "psr/log": "^1.1|^2|^3",
         "symfony/cache-contracts": "^1.1.7|^2|^3",
