@@ -31,7 +31,7 @@ class EarlyExpirationDispatcher
     {
         $this->bus = $bus;
         $this->reverseContainer = $reverseContainer;
-        $this->callbackWrapper = null === $callbackWrapper || $callbackWrapper instanceof \Closure ? $callbackWrapper : \Closure::fromCallable($callbackWrapper);
+        $this->callbackWrapper = null === $callbackWrapper ? null : $callbackWrapper(...);
     }
 
     public function __invoke(callable $callback, CacheItem $item, bool &$save, AdapterInterface $pool, \Closure $setMetadata, LoggerInterface $logger = null)
