@@ -62,7 +62,7 @@ class EarlyExpirationHandlerTest extends TestCase
 
         $handler($msg);
 
-        $this->assertSame(123, $pool->get('foo', [$this, 'fail'], 0.0, $metadata));
+        $this->assertSame(123, $pool->get('foo', $this->fail(...), 0.0, $metadata));
 
         $this->assertGreaterThan(25, $metadata['ctime']);
         $this->assertGreaterThan(time(), $metadata['expiry']);
